[Addressing comments # 4]
This commit addresses comments on previous commit and makes the following changes

1. changes usage of __str__() to str()
2. changes error = e.__str__() to error = e.message to be consistent with the message parameter
3. adds a link to the docs for the error message (for both add docs & search) using marqo_docs.py module. A seperate PR has been raised to add zeroMagnitudeVectorError related documentation in marqo docs (https://github.com/marqo-ai/marqodocs/pull/327)
4. Moves API tests to marqo-api-test repo (PR: https://github.com/marqo-ai/marqo-api-tests/pull/97)
5. Deletes test_custom_vector_index_normalize_embeddings_true.py file and moves test case to test_custom_vector_field.py
6. Fixes exception message related convention (https://github.com/marqo-ai/marqo/pull/970#discussion_r1776416433)
