+++ src/marqo/core/search/hybrid_search.py
@@ -90,14 +90,14 @@ class HybridSearch:
                 marqo_index_version < constants.MARQO_STRUCTURED_HYBRID_SEARCH_MINIMUM_VERSION:
             raise core_exceptions.UnsupportedFeatureError(
                 f"Hybrid search is only supported for Marqo structured indexes created with Marqo "
-                f"{constants.MARQO_STRUCTURED_HYBRID_SEARCH_MINIMUM_VERSION.__str__()} or later. "
+                f"{str(constants.MARQO_STRUCTURED_HYBRID_SEARCH_MINIMUM_VERSION)} or later. "
                 f"This index was created with Marqo {marqo_index_version}."
             )
         elif isinstance(marqo_index, UnstructuredMarqoIndex) and \
                 marqo_index_version < constants.MARQO_UNSTRUCTURED_HYBRID_SEARCH_MINIMUM_VERSION:
             raise core_exceptions.UnsupportedFeatureError(
                 f"Hybrid search is only supported for Marqo unstructured indexes created with Marqo "
-                f"{constants.MARQO_UNSTRUCTURED_HYBRID_SEARCH_MINIMUM_VERSION.__str__()} or later. "
+                f"{str(constants.MARQO_UNSTRUCTURED_HYBRID_SEARCH_MINIMUM_VERSION)} or later. "
                 f"This index was created with Marqo {marqo_index_version}."
             )
