+++ src/marqo/version.py
@@ -1,4 +1,4 @@
-__version__ = "2.12.1"
+__version__ = "2.13.0"
 
 def get_version() -> str:
     return f"{__version__}"
