+++ tests/tensor_search/integ_tests/test_search_unstructured.py
@@ -1115,7 +1115,7 @@ class TestSearchUnstructured(MarqoTestCase):
             )
         )
 
-        alright_queries = [{"v ": 1.2}, {"d ": 0}, {"vf": -1}]
+        alright_queries = [{"v ": 1.2}, {"vf": -1}]
         for q in alright_queries:
             with self.subTest(f"query={alright_queries}"):
                 tensor_search.search(
